requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
SQLAlchemy==2.0.27
typing_extensions==4.10.0
Werkzeug==3.0.1
pillow == 10.2.0
cloudinary == 1.39.0